apex_harness/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
tqdm>=4.66.0
click>=8.1.0
rich>=13.7.0
jinja2>=3.1.0
python-dotenv>=1.0.0
flask>=3.0.0
flask-socketio>=5.3.6
pywebview>=5.0.0
pyinstaller>=6.0.0
